Fix a few coverity complaints
All minor:
- execdir does not need initialization, it's not used until written to
- 'newest' could be NULL
- zalloc(-1) confuses coverity
- 't' is never used in that test

+++ test/litest-selftest.c
@@ -350,7 +350,7 @@ END_TEST
 
 START_TEST(zalloc_overflow)
 {
-	zalloc(-1);
+	zalloc((size_t)-1);
 }
 END_TEST
